Task A5 (инкремент 4): устранение CS-предупреждений + zero-NEW-warning baseline (66 to 0)
CS0414/0675/8618/8765/8600-8625 исправлены; Puncher OpenWriter refactor; baseline 0 CS (допущено только уменьшение). build 0 CS, check PASS exit 0, fast 3/3 OK, Python 40 OK.

+++ src/besm6.net/Asm/Assembler.cs
@@ -128,7 +128,7 @@ namespace Besm6.Asm
                 }
             }
 
-            result = (ulong)reg << 20 | (ulong)opcode << 12 | (ulong)addr;
+            result = (uint)reg << 20 | (uint)opcode << 12 | (uint)addr;
             return src;
         }
